feat(publish-tag): publish the helm chart on tag push
This is an interim solution for this next release until we sort out what the automated
procedure will be like.
We'll keep the the versions and images pointing to the v2.0.0 release, so the chart will not
be working from the git repo as is, without overriding the image tags.

When a new tag is pushed (eg: 2.0.0-alpha.2), we publish a helm chart with that semver for
the appVersion and for the version. Exception: if the version has already increased, we'll
simply bump the version.

A test has been added which shows how this works.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>